split NodeTrait to NodeMethodsTrait and NodePropertiesTrait
